+++ CHANGELOG.md
@@ -4,13 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
 
 
+## [1.3.0] - 2022-02-26
+### Added
+- New `.warning()` alert modal variant.
+- New `sanitizer` modal configuration option which can be used to sanitize the modal markup string before parsing and inserting into the DOM ([docs](https://erobertson42.github.io/bootstrap-modbox/docs.html#options-sanitizer)).
+- New `sanitizer` input configuration option for `.prompt()` modals which can be used to sanitize user input before being passed to the Promise resolve function ([docs](https://erobertson42.github.io/bootstrap-modbox/docs.html#options-input)).
+
+### Changed
+- Safari has now supported [private class methods](https://caniuse.com/mdn-javascript_classes_private_class_methods) for a few versions, so removed the Grunt task that converted them into private field functions (the compatibility file still exists for older browser versions).
+- Updated documentation to reflect changes.
+
+
 ## [1.2.1] - 2021-10-23
 ### Added
 - New `message` modal configuration option as an alternative to `body` ([docs](https://erobertson42.github.io/bootstrap-modbox/docs.html#options-message)).
 
 ### Changed
 - The 2nd argument for the `addButton()` instance method has been renamed from `appendStart` to `swapOrder` to more accurately reflect its behavior.
-- Updated documentation.
+- Updated documentation to reflect changes.
 
 
 ## [1.2.0] - 2021-10-08